+++ components/primitive/injector/demo/article/ArticleStyles.js
@@ -1,0 +1,33 @@
+import PropTypes from 'prop-types'
+
+import {H2, Paragraph, Article, Box} from '@s-ui/documentation-library'
+
+import Injector from 'components/primitive/injector/src/index.js'
+import Child from '../Child.js'
+
+const ArticleStyles = ({className}) => {
+  return (
+    <Article className={className}>
+      <H2>styles</H2>
+      <Paragraph>
+        Injector will combine existing children styles rule attribute over the
+        given styles prop to all of them.
+      </Paragraph>
+      <Box style={{paddingLeft: 0, paddingRight: 0}}>
+        <Injector style={{fontSize: '2em', fontColor: 'red'}}>
+          <Child
+            name="result"
+            style={{fontSize: '3em', fontFamily: 'Helvetica'}}
+          />
+        </Injector>
+      </Box>
+    </Article>
+  )
+}
+ArticleStyles.displayName = 'ArticleStyles'
+
+ArticleStyles.propTypes = {
+  className: PropTypes.string
+}
+
+export default ArticleStyles